refactor: extract shared preview block components for Mermaid and PlantUML
Move duplicate _BlockColors, _TabButton, _TextAction, loading/error views
to the shared tabbed_preview_block.dart. No behavioral changes.
